Data Analysis/requirements.txt
requests==2.32.3
python-dotenv==1.0.1
anthropic==0.40.0
jinja2==3.1.4
weasyprint==62.3
openpyxl==3.1.5
